requirements.txt
async-generator==1.10
attrs==21.2.0
Automat==20.2.0
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.7
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
numpy==1.21.4
outcome==1.1.0
pandas==1.3.4
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
queuelib==1.6.1
requests==2.26.0
Scrapy==2.5.0
scrapy-selenium==0.0.7
selenium==4.0.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
Twisted==21.2.0
urllib3==1.26.7
w3lib==1.22.0
wsproto==1.0.0
zope.interface==5.4.0
